More update to make a generalized createAll

+++ Model/t1BuildInc/Cannelloni.h
@@ -3,7 +3,7 @@
  
  * File:   t1BuildInc/Cannelloni.h
  *
- * Copyright (c) 2004-2018 by Stuart Ansell
+ * Copyright (c) 2004-2019 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -111,9 +111,9 @@ class Cannelloni : public constructSystem::TargetBase
 
   void addProtonLine(Simulation&,const attachSystem::FixedComp&,
 		     const long int);
-  void createAll(Simulation&,const attachSystem::FixedComp&,const int);
+  void createAll(Simulation&,const attachSystem::FixedComp&,
+		 const long int);
   
-
 };
 
 }
