+++ Model/t1BuildInc/SideCoolTarget.h
@@ -91,7 +91,7 @@ class SideCoolTarget :
   void addProtonLine(Simulation&,	 
 		     const attachSystem::FixedComp& refFC,
 		     const long int index);
-  void createAll(Simulation&,const attachSystem::FixedComp&,const int);
+  void createAll(Simulation&,const attachSystem::FixedComp&,const long int);
   
 
 };
